% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_transparency.R
\name{add_transparency}
\alias{add_transparency}
\title{Add transparency to a color}
\usage{
add_transparency(color, trans)
}
\arguments{
\item{color}{character hex color that you want to add transparency to}

\item{trans}{integer value between 0 and 255, with 0 being fully transparent and 255 being fully visible}
}
\value{
A character hex value for the transparent version of your color
}
\description{
Add transparency to a color
}
\examples{
color_in <- "#FF5555"
color_alpha <- add_transparency(color = color_in, trans = 255/2)
plot(1:3, rep(5,times = 3), col = color_alpha, pch = 19)
points(1:3, rep(3, times = 3),col = color_in, pch = 19)
}
