% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthen_pal.R
\name{lengthen_pal}
\alias{lengthen_pal}
\title{Lengthen color palette}
\usage{
lengthen_pal(newlength = 10, shortpal)
}
\arguments{
\item{newlength}{numeric. Length you need your new palette to be.}

\item{shortpal}{A character vector of color names or hex codes; your "starter palette".}
}
\value{
a new character vector containing a new palette of length newlength
}
\description{
Lengthen color palette
}
\examples{
lengthen_pal(shortpal = c('red','yellow','blue'))
}
